display the slug on the email setup page
